+++ CHANGELOG.md
@@ -1,7 +1,14 @@
-# 0.3.0
+# master (will be 0.3.0)
 
+* Added argument `output_buffer_size` to `multicore.Pool.imap_batches()`
+  and `multicore.Pool.imap_batches_unordered()` to control the maximum number
+  of batches in the background augmentation pipeline (allows to limit
+  maximum RAM demands).
 * Increased `max_distance` thresholds for `almost_equals()`, `exterior_almost_equals()` and `coords_almost_equals()` in `Polygon` and `LineString` from `1e-6` to `1e-4`.
-  This should fix false-negative problems related to float inaccuracies. 
+  This should fix false-negative problems related to float inaccuracies.
+
+## Fixes
+ 
 * Fixed an issue with `Polygon.clip_out_of_image()`,
   which would lead to exceptions if a polygon had overlap with an image,
   but not a single one of its points was inside that image plane. 
@@ -48,7 +55,6 @@ Changes related to the new modules:
 * Refactored `__init__()` of `PolygonsOnImage`, `BoundingBoxesOnImage`, `KeypointsOnImage` to make use of `imgaug.augmentables.utils.normalize_shape()`.
 * Refactored `KeypointsOnImage.on()` to use `imgaug.augmentables.utils.normalize_shape()`.
 * Refactored `Keypoint.project()` to use `imgaug.augmentables.utils.project_coords()`.
-
 
 ## Polygon Augmentation